refactor: change the hierarchy between test and testing
